chore: move path parsing into separate module
